Add Analyzer to check that Commands are not structs
